Put Date Truncation in its own file and removed unnessecary blank line
